+++ R/get_problems.R
@@ -4,10 +4,12 @@
 #' self-self, best, and 2nd best distances, and summarize the results
 #' in a data frame.
 #'
-#' @param d A distance matrix
+#' @param d A distance or similarity matrix
 #' @param dimension Whether to determine the best distances within rows or columns
-#' @param get_min If TRUE, get the minimum; if FALSE, get the maximum
+#' @param get_min If TRUE, get the minimum (for a distance matrix);
+#'     if FALSE, get the maximum (for a similarity matrix)
 #' @param subset Whether to return just the rows with potential problems, or all of the rows.
+#' @param threshold If `subset="problems"`, the threshold on the difference between the self and best distances.
 #'
 #' @return A data frame containing individual ID, distance to self,
 #'     best distance and corresponding individual, 2nd best distance
@@ -18,7 +20,7 @@
 #' @export
 get_problems <-
     function(d, dimension=c("row", "column"), get_min=TRUE,
-             subset=c("problems", "all"))
+             subset=c("problems", "all"), threshold=0)
 
 {
     dimension <- match.arg(dimension)
@@ -35,10 +37,17 @@ get_problems <-
     else result <- result[order(result$self - result$best),,drop=FALSE]
 
     subset <- match.arg(subset)
-    if(subset=="all") return(result)
 
-    if(get_min)
-        return(result[is.na(result$self) | is.na(result$best) | result$best < result$self,])
-    else
-        return(result[is.na(result$self) | is.na(result$best) | result$best > result$self,])
+    if(subset == "problems") {
+        if(get_min) {
+            result <- result[(is.na(result$self) & !is.na(result$best)) |
+                             (!is.na(result$best) & !is.na(result$self) & result$best < result$self - threshold),]
+        } else {
+            result <- result[(is.na(result$self) & !is.na(result$best)) |
+                             (!is.na(result$best) & !is.na(result$self) & result$best > result$self + threshold),]
+        }
+    }
+
+    rownames(result) <- seq_len(nrow(result))
+    result
 }